Create TeamChild module
